+++ packages/as-chacha20poly1305/src/index.ts
@@ -12,17 +12,6 @@ const keyUint8Array = new Uint8Array(ctx.memory.buffer, wasmKeyValue, ctx.KEY_LE
 const counterUint8Array = new Uint8Array(ctx.memory.buffer, wasmCounterValue, ctx.COUNTER_LENGTH);
 const debugUint8Array = new Int32Array(ctx.memory.buffer, ctx.debug.value, 64);
 
-export function numberOperation(
-  x0: number, x1: number, x2: number,
-  x3: number, x4: number, x5: number,
-  x6: number, x7: number, x8: number,
-  x9: number, x10: number, x11: number,
-  x12: number, x13: number, x14: number,
-  x15: number
-  ): number {
-  return ctx.numberOperation(x0, x1, x2, x3, x4, x5, x6, x7, x8, x9, x10, x11, x12, x13, x14, x15);
-}
-
 export function streamXOR(key: Uint8Array, nonce: Uint8Array, src: Uint8Array): Uint8Array {
   // We only support 256-bit keys.
   if (key.length != ctx.KEY_LENGTH) {
@@ -48,8 +37,6 @@ export function streamXOR(key: Uint8Array, nonce: Uint8Array, src: Uint8Array): 
     const dataLength = ctx.streamXORUpdate(length);
     output.set(outputUint8Array.subarray(0, dataLength), start);
   }
-
-  // loop here
 
   return output;
 }
