+++ packages/as-chacha20poly1305/src/wasm.ts
@@ -27,14 +27,6 @@ export interface WasmContext {
   };
 
   streamXORUpdate(dataLength: number): number;
-  numberOperation(
-    x0: number, x1: number, x2: number,
-    x3: number, x4: number, x5: number,
-    x6: number, x7: number, x8: number,
-    x9: number, x10: number, x11: number,
-    x12: number, x13: number, x14: number,
-    x15: number
-    ): number;
 }
 
 const importObj = {
